Changed the code so the flow makes more sense
