+++ app/src/main/java/com/example/volunteeringapp/Homepage.java
@@ -4,36 +4,51 @@ import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
 
+import android.content.Intent;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
+import android.view.View;
 import android.widget.Toast;
 
-public class Account extends AppCompatActivity {
+public class Homepage extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_account);
+        setContentView(R.layout.activity_homepage);
         Toolbar myToolbar = (Toolbar) findViewById(R.id.toolbar);
         setSupportActionBar(myToolbar);
         getSupportActionBar().setDisplayShowTitleEnabled(false);
+
     }
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.account_menu, menu);
+        getMenuInflater().inflate(R.menu.main_menu, menu);
         return true;
     }
-
+//
     @Override
     public boolean onOptionsItemSelected(@NonNull MenuItem item) {
         switch (item.getItemId()) {
-            case R.id.editButton:
-                Toast.makeText(getApplicationContext(), "Edit Profile", Toast.LENGTH_LONG).show();
+            case R.id.search:
+                Toast.makeText(getApplicationContext(), "Search", Toast.LENGTH_LONG).show();
                 return true;
             default:
                 return super.onOptionsItemSelected(item);
         }
     }
+
+    public boolean navigateToEventDetails(View view){
+        Intent eventDetails = new Intent (this, EventDetails.class);
+        startActivityForResult(eventDetails, 1);
+        return true;
+    }
+
+//    public boolean navigateToAccount(View view){
+//        Intent eventDetails = new Intent(this, Account.class);
+//        startActivityForResult(eventDetails, 1);
+//        return true;
+//    }
 }
