+++ app/src/main/java/com/example/volunteeringapp/Login.java
@@ -1,0 +1,22 @@
+package com.example.volunteeringapp;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+
+public class Login extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_login);
+    }
+
+    public boolean navigateToHome(View view){
+        Intent home = new Intent (this, Homepage.class);
+        startActivityForResult(home, 1);
+        return true;
+    }
+}
