+++ app/src/main/java/com/example/volunteeringapp/MainActivity.java
@@ -17,38 +17,17 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        Toolbar myToolbar = (Toolbar) findViewById(R.id.toolbar);
-        setSupportActionBar(myToolbar);
-        getSupportActionBar().setDisplayShowTitleEnabled(false);
-
     }
 
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.main_menu, menu);
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(@NonNull MenuItem item) {
-        switch (item.getItemId()) {
-            case R.id.search:
-                Toast.makeText(getApplicationContext(), "Search", Toast.LENGTH_LONG).show();
-                return true;
-            default:
-                return super.onOptionsItemSelected(item);
-        }
-    }
-
-    public boolean navigateToEventDetails(View view){
-        Intent eventDetails = new Intent (this, EventDetails.class);
+    public boolean navigateToLogin(View view){
+        Intent eventDetails = new Intent (this, Login.class);
         startActivityForResult(eventDetails, 1);
         return true;
     }
 
-    public boolean navigateToAccount(View view){
-        Intent eventDetails = new Intent (this, Account.class);
-        startActivityForResult(eventDetails, 1);
+    public boolean navigateToSignup(View view){
+        Intent signup = new Intent (this, Signup.class);
+        startActivityForResult(signup, 1);
         return true;
     }
