+++ app/src/main/res/values/themes.xml
@@ -3,7 +3,7 @@
     <style name="Theme.VolunteeringApp" parent="Theme.MaterialComponents.DayNight.DarkActionBar">
         <!-- Primary brand color. -->
         <item name="colorPrimary">#005280</item>
-        <item name="colorPrimaryVariant">#005280</item>
+        <item name="colorPrimaryVariant">#013653</item>
         <item name="colorOnPrimary">@color/white</item>
         <!-- Secondary brand color. -->
         <item name="colorSecondary">#F4511E</item>
